Add code for a grid size based on user input
I changed the original code for the first grid.
Made it so that a row is added to the container
and each box is added to that row.
That repeats until the number is reached

Also moved the event listeners out of the
for loops for better performance.

Removed the button creation and put it
in the html file for convenience.

Used `container.innerHTML = ''` to clear the
baseGrid and make room for the new grid.